#78 refactor: 토큰관련 enum의 패키지 위치 변경
