+++ src/main/java/com/pettoyou/server/domains/auth/enums/TokenType.java
@@ -1,4 +1,4 @@
-package com.pettoyou.server.config.jwt.util;
+package com.pettoyou.server.domains.auth.enums;
 
 public enum TokenType {
     ACCESS_TOKEN, REFRESH_TOKEN
