+++ src/main/java/com/pettoyou/server/domains/auth/enums/TokenUserType.java
@@ -1,4 +1,4 @@
-package com.pettoyou.server.config.jwt.util;
+package com.pettoyou.server.domains.auth.enums;
 
 public enum TokenUserType {
     MEMBER_TOKEN, HOSPITAL_ADMIN_TOKEN
